pyproject.toml
[project]
name = "pprp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-allauth[socialaccount]>=65.14.3",
    "django-map-widgets>=0.5.1",
    "dotenv>=0.9.9",
    "pillow>=12.1.1",
    "psycopg2>=2.9.11",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
]
